adding custom authentication & adding archived bill list with ability to archive unarchive functionality

+++ bills/billsapi/serializers.py
@@ -5,7 +5,7 @@ from .models import Bills, BillsItems
 class BillsSerializer(serializers.ModelSerializer):
     created_at = serializers.SerializerMethodField()
     user = serializers.StringRelatedField(read_only=True)
-    total = serializers.IntegerField(read_only=True)
+    total = serializers.FloatField(read_only=True)
 
     class Meta:
         model = Bills
